create remote task interface

fix test


change client to remote task


refactor the interface


set gcpbatchclient as client for batch service staticaly


fix copyright


feat: Add unit tests for BatchService

Adds a new test file, , to provide comprehensive unit
test coverage for the . This includes testing the
 and  methods,
ensuring the batch job creation logic is working as expected.

Also, refactors the batch module to introduce a
and moves the GCP Batch implementation to its own module. This improves
modularity and makes the batch service more extensible.

enrich docstrings


docs: Enrich docstrings for batch module

Enriches the docstrings for the batch module, with a focus on  to
explain the use of GCP Batch for running granular tasks with VM-level
isolation for untrusted code. Also adds docstrings to ,
, , and .

Fixes linting errors in .


fix: Add missing docstrings and fix lint errors

Adds missing docstrings to functions in the  file to resolve linting errors. This improves code clarity and ensures conformance with project coding standards.

refactor: Move batch service tests and enrich docstrings

Moves tests for the batch service from  to  to better align with the new module structure. Deletes the now-redundant .

Also enriches docstrings across the batch module, including , , , and , to improve code clarity and explain the security model of using GCP Batch for isolated task execution.

Fixes all related linting errors and ensures all tests continue to pass.

move batch service test


fix: Address linting and docstring issues in run_remote_task.py

Adds a module docstring, ensures consistent return statements in ,
disables unused argument warnings for , removes the unused  import,
and adds the missing license header to .
